chore(app): gen midgard client and use server component for explore page (#12)
Ninerealms exposes a openapi spec. Using it makes it easier to use
generated types. Fortunately the spec appears to be well formed, so
it's a pretty easy win. This commit also adds a client gen script
to package.json.

In order for the generator to work, I've added pnpm. npm was bombing out
installing the openapi lib for some reason.

In addition to updating the explore page to use the generated ts types,
I refactored the explore page to be a server component. This makes the
page appear to load faster and removes the need to have a loading text.

Unfortunately, vanaheimex does not have a spec or at least i could not
find any documentation mentioning one. Nonetheless, I added some types
so that the compiler is happy parsing the untyped json response.

This also fixes build errors from bad types

Co-authored-by: desamtralized <sambarbosaa@gmail.com>